+++ WebApi/Models/DTO/DTOBegrotingVoorstel.cs
@@ -9,7 +9,7 @@ namespace WebApi.Models.DTO
     {
         public int Id { get; set; }
         public int aantalStemmen { get; set; }
-        public int totaal { get; set; }
+        public float totaal { get; set; }
         public string samenvatting { get; set; }
         public string beschrijving { get; set; }
         public List<DTOBudgetWijziging> budgetWijzigingen { get; set; }
